+++ src/test/java/one/util/streamex/StreamExTest.java
@@ -109,6 +109,8 @@ public class StreamExTest {
 
         assertEquals(asList("a", "b"), StreamEx.of(asList("a", "b").spliterator()).toList());
         assertEquals(asList("a", "b"), StreamEx.of(asList("a", "b").iterator()).toList());
+        assertEquals(asList(), StreamEx.of(asList().iterator()).toList());
+        assertEquals(asList(), StreamEx.of(asList().iterator()).parallel().toList());
         assertEquals(asList("a", "b"), StreamEx.of(new Vector<>(asList("a", "b")).elements()).toList());
     }
 
@@ -140,6 +142,8 @@ public class StreamExTest {
 
         assertEquals(expectedSet, StreamEx.ofLines(new StringReader(input)).parallel().skipOrdered(1).toSet());
         assertEquals(expectedSet, StreamEx.ofLines(new StringReader(input)).skipOrdered(1).parallel().toSet());
+        
+        assertFalse(StreamEx.ofLines(new StringReader(input)).spliterator().getClass().getSimpleName().endsWith("IteratorSpliterator"));
     }
 
     @Test(expected = IllegalArgumentException.class)
@@ -1513,6 +1517,10 @@ public class StreamExTest {
 
     @Test
     public void testSplit() {
+        assertFalse(StreamEx.split("str", "abcd").spliterator().getClass().getSimpleName()
+                .endsWith("IteratorSpliterator"));
+        assertFalse(StreamEx.split("str", Pattern.compile("abcd")).spliterator().getClass().getSimpleName()
+                .endsWith("IteratorSpliterator"));
         streamEx(() -> StreamEx.split("", "abcd"), s -> assertEquals(1, s.get().count()));
         streamEx(() -> StreamEx.split("", Pattern.compile("abcd")), s -> assertEquals(1, s.get().count()));
         streamEx(() -> StreamEx.split("ab.cd...", '.'), s -> assertEquals("ab|cd", s.get().joining("|")));