+++ CHANGES.md
@@ -3,6 +3,8 @@
 ### 0.5.2
 * [#19] Optimized: pairMap and forPairs may work faster now, especially in the presence of upstream intermediate operations.
 * [#42] Added: `EntryStream.ofTree` methods to stream the tree-like structure tracking nodes depth.
+* [#46] Optimized: parallel performance of all `of(Iterator)` methods.
+* [#3] Optimized: parallel performance of `StreamEx.ofLines` as well as `StreamEx.split`
 
 ### 0.5.1